backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
numpy==1.26.4
requests==2.31.0
